Hotfix to allow for compilation

+++ src/include/a_tensor.hpp
@@ -79,7 +79,7 @@ class Tensor {
   /// @brief Fills the tensor with a given value.
   /// @param value The value to fill the tensor with.
   void fill(T value) {
-    for (int i = 0; i < this->size; i++) {
+    for (uint64_t i = 0; i < this->size; i++) {
       (*this)[i] = value;
     }
   }
